Rename test package `framework` to `testFramework`
